requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
python-jose
pydantic
email-validator